validate sample rate, rewrite shouldSample method
Previously, if you had a sample rate of 0.01, then
Math.round(1 / 0.01) = 100 and shouldSample method was doing

ThreadLocalRandom.current().nextInt(samplingRate) == 0

which picked a number from 0 (inclusive) to 100 (exclusive).

However, if we set sample rate to, for example, 0.75 to say that
75% should be sampled, then Math.round(1 / 0.75) = 1.33 and rounded it is 1.

ThreadLocalRandom.current().nextInt(1) == 0

will be always true. That means what basically from some sample rate which rounds to 1
we lose the probability.

The current approach works with floats and it is rewritten to

ThreadLocalRandom.current().nextFloat() < samplingRate

nextFloat() gives values between zero (inclusive) and one (exclusive).
If we set sampling rate to 1 then it will be always true. If we set it to 0.01 that will be 1%.
If we set it to 0.75 that will be 75%, without losing any accuracy.

+++ src/java/org/apache/cassandra/db/compression/ICompressionDictionaryTrainer.java
@@ -125,10 +125,10 @@ public interface ICompressionDictionaryTrainer extends AutoCloseable
     /**
      * Updates the sampling rate for this trainer.
      *
-     * @param newSamplingRate the new sampling rate. For exmaple, 1 = sample every time (100%),
-     *                        2 = expect sample 1/2 of data (50%), n = expect sample 1/n of data
+     * @param newSamplingRate the new sampling rate. For exmaple, 0.01 - sample 1% of data,
+     *                        1 = sample every time (100%), 0.5 - sample 50% of data.
      */
-    void updateSamplingRate(int newSamplingRate);
+    void updateSamplingRate(float newSamplingRate);
 
     /**
      * Factory method to create appropriate trainer based on compression parameters.
